Bug 1303054 - use dynamic cache size with SkiaGL on desktop. r=milan
MozReview-Commit-ID: 3Hhsa1SNcXw

UltraBlame original commit: 2d0944de7efb8ce438e160d622aa2a266c5fb3f4

+++ gfx/thebes/gfxPlatform.h
@@ -770,7 +770,7 @@ protected:
     
     int32_t mWordCacheMaxEntries;
 
-    uint32_t mTotalSystemMemory;
+    uint64_t mTotalSystemMemory;
 
     
     RefPtr<mozilla::gfx::VsyncSource> mVsyncSource;
